+++ components/card/stories/template.js
@@ -95,6 +95,7 @@ export const Template = ({
               class=${classMap({ [`${rootClass}-coverPhoto`]: true })}
               style=${styleMap({ "background-image": `url(${image})` })}
             ></div>
+            <hr class="spectrum-Divider spectrum-Divider--sizeS spectrum-Card-divider">
           `
         )
       )}
@@ -178,7 +179,7 @@ export const Template = ({
           </div>`
       )}
       ${when(
-        footer,
+        footer && !isQuiet,
         () => html`
           <div
             class=${classMap({