+++ components/rating/stories/rating.stories.js
@@ -5,7 +5,12 @@ import { RatingGroup } from "./rating.test.js";
 import { Template } from "./template.js";
 
 /**
- * A rating element is used to display or collect a user's rating of an item as represented by a number of stars.
+ * The rating component is used to display or collect a user's rating of an item as represented by a number of stars.
+ * 
+ * ### Usage notes
+ * - All active stars have the class `is-selected` applied.
+ * - The current value (the last active star) has the class `is-currentValue` applied.
+ * - When the rating receives focus, the class `is-focused` should be added to the component's root element (`.spectrum-Rating`).
  */
 export default {
 	title: "Rating",
@@ -17,15 +22,18 @@ export default {
 		isReadOnly,
 		max: {
 			name: "Maximum value",
+			description: "The total number of stars. Star ratings should always have 5 available stars. This shouldn't be increased or decreased to fit various containers.",
 			type: { name: "number" },
 			table: {
 				type: { summary: "number" },
 				category: "Content",
+				disable: true,
 			},
 			control: { type: "number" },
 		},
 		value: {
 			name: "Value",
+			description: "The current rating, represented by the number of selected stars.",
 			type: { name: "number" },
 			table: {
 				type: { summary: "number" },
@@ -38,6 +46,8 @@ export default {
 	args: {
 		rootClass: "spectrum-Rating",
 		isDisabled: false,
+		isEmphasized: false,
+		isReadOnly: false,
 		max: 5,
 		value: 3,
 	},
@@ -46,6 +56,11 @@ export default {
 	},
 };
 
+/**
+ * A initial value of three is used for the following examples, to demonstrate both active and inactive stars.
+ * When hovering over a rating component that has a previously entered value, an underline appears under the
+ * current selection to provide context.
+ */
 export const Default = RatingGroup.bind({});
 Default.args = {};
 
@@ -60,7 +75,12 @@ WithForcedColors.parameters = {
 };
 
 // ********* DOCS ONLY ********* //
+
+/**
+ * A non-interactive rating.
+ */
 export const ReadOnly = Template.bind({});
+ReadOnly.storyName = "Read-only";
 ReadOnly.tags = ["!dev"];
 ReadOnly.args = {
 	isReadOnly: true,
@@ -79,6 +99,7 @@ Emphasized.parameters = {
 };
 
 export const ReadOnlyEmphasized = Template.bind({});
+ReadOnlyEmphasized.storyName = "Read-only, emphasized";
 ReadOnlyEmphasized.tags = ["!dev"];
 ReadOnlyEmphasized.args = {
 	isEmphasized: true,