+++ components/sidenav/stories/sidenav.test.js
@@ -23,28 +23,28 @@ export const SideNavGroup = Variants({
 					levelTwoItems: [
 						{
 							id: "section2.1",
-							title: "Section title 1",
+							title: "Section title A",
 							link: "#",
 						},
 						{
 							id: "section2.2",
-							title: "Section title 2: The long title that wraps to the next line",
+							title: "Section title B: The long title that wraps to the next line",
 							link: "#",
 							levelThreeItems: [
 								{
 									id: "section3.1",
-									title: "Section title 1",
+									title: "Section title X",
 									link: "#",
 								},
 								{
 									id: "section3.2",
-									title: "Section title 2: Another long title that wraps to the next line",
+									title: "Section title Y: this menu item is set as current / selected",
 									link: "#",
 									isSelected: true,
 								},
 								{
 									id: "section3.3",
-									title: "Section title 3",
+									title: "Section title Z",
 									link: "#",
 								},
 							]
@@ -56,7 +56,7 @@ export const SideNavGroup = Variants({
 					title: "Section title 3",
 					link: "#",
 				},
-			]
+			],
 		},
 		{
 			testHeading: "With headings",
@@ -73,17 +73,29 @@ export const SideNavGroup = Variants({
 					levelTwoItems: [
 						{
 							id: "section2.1",
-							title: "Section 1",
+							title: "Section 2",
 							link: "#",
 						},
 						{
 							id: "section2.2",
-							title: "Section 2",
+							title: "Section 3",
+							link: "#",
+						},
+					]
+				},
+				{
+					id: "section3",
+					heading: "Heading 2",
+					link: "#",
+					levelTwoItems: [
+						{
+							id: "section3.1",
+							title: "Section 4",
 							link: "#",
 						},
 					]
 				}
-			]
+			],
 		},
 		{
 			testHeading: "With icon",