+++ components/stepper/CHANGELOG.md
@@ -203,6 +203,15 @@
   - @spectrum-css/tokens@15.0.0-s2-foundations.0
   - @spectrum-css/icon@8.0.0-s2-foundations.0
 
+## 6.1.4
+
+### Patch Changes
+
+- [#3012](https://github.com/adobe/spectrum-css/pull/3012) [`78c98bf`](https://github.com/adobe/spectrum-css/commit/78c98bf15bfc82656acd14a4a22fc537f722b60d) Thanks [@cdransf](https://github.com/cdransf)! - Modify the stepper and textfield css to allow the .is-disabled state and class to properly disable the border behavior that would occur when a disabled stepper was hovered. It also leverages the outline state for focused components to align with the state currently used by button components.
+
+- Updated dependencies [[`78c98bf`](https://github.com/adobe/spectrum-css/commit/78c98bf15bfc82656acd14a4a22fc537f722b60d)]:
+  - @spectrum-css/textfield@7.2.2
+
 ## 6.1.3
 
 ### Patch Changes